fixed some tests;
