+++ weave/trace/vals.py
@@ -94,7 +94,7 @@ def unwrap(val: Any) -> Any:
     if isinstance(val, Traceable):
         return val.unwrap()
     elif isinstance(val, ObjectRecord):
-        return unwrap(val.unwrap_one_level())
+        return val.unwrap()
     elif isinstance(val, dict):
         return {k: unwrap(v) for k, v in val.items()}
     elif isinstance(val, list):
